+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.6.2
 - [Tech Preview] Add sample code in InstanaAgentExample app on how to enable crash reporting to Instana backend
+- Once in slow send mode, periodically resend 1 beacon. Once out of slow send mode, flush all beacons immediately.
 
 ## 1.6.1
 - Fix crash caused by appendMetaData() inside InstanaProperties class